% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{analyze_data}
\alias{analyze_data}
\title{Analyze Data}
\usage{
analyze_data(data, print_results = TRUE)
}
\arguments{
\item{data}{A data frame containing the data to be analyzed.}

\item{print_results}{Logical value indicating whether to print the results. Default is TRUE.}
}
\value{
A data frame with the analyzed statistics.
}
\description{
This function analyzes the given data and returns various statistics.
}
