% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_columnnames.R
\name{sort_columnnames}
\alias{sort_columnnames}
\title{Sort column names of a data frame}
\usage{
sort_columnnames(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame with column names sorted in alphabetical order.
}
\description{
This function sorts the column names of a data frame in alphabetical order.
}
