% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_NA_with_y.R
\name{replace_NA_with_y}
\alias{replace_NA_with_y}
\title{Replace NA with y}
\usage{
replace_NA_with_y(data, x, y)
}
\arguments{
\item{data}{a data frame}

\item{x}{a (quoted) variable name with missing values.}

\item{y}{a (quoted) variable name with values that can be imputed in x.}
}
\value{
the dataframe with the imputed x variable. The y variable is removed from the output.
}
\description{
Een functie om missende waarden van x op te vullen met de waarde van y voor
diezelfde rij. Deze functie werkt zoals coalesce, maar hier kan een data frame
als imput gegeven worden.
}
\concept{missing data functies}
\concept{vector calculations}
