% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strict_left_join.R
\name{strict_left_join}
\alias{strict_left_join}
\title{strict_left_join}
\usage{
strict_left_join(x, y, by = NULL, ...)
}
\arguments{
\item{x}{data frame x (left)}

\item{y}{data frame y (right)}

\item{by}{unquoted variable names to join.}

\item{...}{Pass further arguments to dplyr::left_join}
}
\value{
merged data frame
}
\description{
A wrapper around dplyr's left_join, with an error message if there
duplicate values are in the matching fields in y. This will double
of rows. See dplyr::left_join .
}
\examples{
df1 <- tibble::tibble(a = 1:2)
df2 <- tibble::tibble(a = 2:3,
                      b = 1:2)
strict_left_join(df1, df2, by = "a")
}
\seealso{
\code{\link[dplyr:join]{left_join}}
}
